+++ lib/models/picture/picture_model.g.dart
@@ -6,13 +6,16 @@ part of 'picture_model.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-PictureModel _$PictureModelFromJson(Map<String, dynamic> json) => PictureModel(
+_$PictureModelImpl _$$PictureModelImplFromJson(Map<String, dynamic> json) =>
+    _$PictureModelImpl(
+      id: (json['id'] as num).toInt(),
       src: json['src'] as String,
       alt: json['alt'] as String,
     );
 
-Map<String, dynamic> _$PictureModelToJson(PictureModel instance) =>
+Map<String, dynamic> _$$PictureModelImplToJson(_$PictureModelImpl instance) =>
     <String, dynamic>{
+      'id': instance.id,
       'src': instance.src,
       'alt': instance.alt,
     };